Move extensions class to extension namespace
