--- LyricsBoard/Core/ComponentModel/Extension/ObservableExtensions.cs
@@ -1,12 +1,0 @@
-﻿//using System;
-
-//namespace LyricsBoard.Core.ComponentModel.Extension
-//{
-//    internal static class ObservableExtensions
-//    {
-//        public static IDisposable WeakSubscribe<T>(this IObservable<T> observable, IObserver<T> observer)
-//        {
-//            return new WeakSubscription<T>(observable, observer);
-//        }
-//    }
-//}
