+++ LyricsBoard/Core/ComponentModel/SimpleDisposable.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace LyricsBoard.Core.ComponentModel
+{
+    internal class SimpleDisposable : IDisposable
+    {
+        private readonly Action disposeAction;
+        public SimpleDisposable(Action disposeAction) => this.disposeAction = disposeAction;
+        public void Dispose() => disposeAction?.Invoke();
+    }
+
+    internal class EmptyDisposable : IDisposable
+    {
+        public void Dispose() { }
+    }
+}
